Refactor + animate amenity
