+++ PolyNavi/Views/MapVC/Map/Annotations/BuildingAnnotation.swift
@@ -1,0 +1,21 @@
+//
+//  BuildingAnnotation.swift
+//  PolyNavi
+//
+//  Created by Andrei Soprachev on 17.01.2022.
+//
+
+import MapKit
+
+class BuildingAnnotation: NSObject, MKAnnotation {
+    @objc dynamic var coordinate: CLLocationCoordinate2D
+    var title: String?
+    var subtitle: String?
+    
+    init(coordinate: CLLocationCoordinate2D, title: String?, subtitle: String?) {
+        self.coordinate = coordinate
+        self.title = title
+        self.subtitle = subtitle
+        super.init()
+    }
+}
